feat(FN-3611): revert to setNoStoreCacheControl
