+++ libs/common/src/middleware/index.ts
@@ -1,3 +1,3 @@
 export * from './feature-flags';
 export * from './create-validation-middleware-for-schema';
-export * from './set-cache-control';
+export * from './set-no-store-cache-control';
